requirements.txt
gevent==21.12.0
jwt==1.3.1
flask_jwt_extended==4.4.1
flask_sqlalchemy==2.5.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
SQLAlchemy==1.4.15
WTForms==3.0.1
flask-wtf==1.0.1
bootstrap4==0.1.0
 dominate-2.6.0 flask_bootstrap-3.3.7.1 visitor-0.1.3